Relocate country list so it can be reused on the server side
